extract type login to separate module
